+++ R/vech.R
@@ -15,7 +15,8 @@
 #' @export
 #'
 #' @examples
-#' x <- matrix(c(1, 2, 2, 4), nrow = 2, byrow = TRUE,
+#' x <- matrix(c(1, 2, 2, 4),
+#'   nrow = 2, byrow = TRUE,
 #'   dimnames = list(1:2, c("Sex", "Smoker"))
 #' )
 #' print(x)
@@ -38,12 +39,13 @@ vech <- function(A, use.Names = TRUE, tol = .Machine$double.eps^0.5) {
   }
 
   symm <- is.symmetric.matrix(A, tol)
-  if (isFALSE(symm))
+  if (isFALSE(symm)) {
     stop(sprintf("% must be a numeric and symmetric matrix for half-vectorization."))
+  }
   # } else { #if symm is TRUE or NA ...
 
   full <- vec(A, use.Names)
-  stack  <- A[lower.tri(A, diag = TRUE)]
+  stack <- A[lower.tri(A, diag = TRUE)]
   vech.A <- full[stack]
 
   return(vech.A)